+++ sdk/core/core-sse/vitest.config.ts
@@ -5,12 +5,15 @@ import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
-    reporters: ["basic", "junit"],
+    reporters: ["verbose", "junit"],
+    fileParallelism: false,
+    testTimeout: 30000,
+    typecheck: {
+      enabled: true,
+    },
+    globalSetup: "test/server/start.mts",
     outputFile: {
       junit: "test-results.xml",
-    },
-    fakeTimers: {
-      toFake: ["setTimeout", "Date"],
     },
     watch: false,
     include: ["test/**/*.spec.ts"],